Add 2 views for select factor enroll/challenge
