+++ playground/mocks/api/v1/idx/challenge/data/factor-required-email-with-options.json
@@ -27,7 +27,7 @@
         "rel": [
           "create-form"
         ],
-        "name": "switch-factor",
+        "name": "select-factor",
         "href": "http://localhost:3000/api/v1/idx/challenge",
         "method": "POST",
         "accepts": "application/vnd.okta.v1+json",
