+++ playground/mocks/api/v1/idx/introspect/index.js
@@ -1,6 +1,6 @@
-// const data1 = require('./data/identify.json');
+const data1 = require('./data/identify.json');
 // const data1 = require('../data/factor-enroll-email.json');
-const data1 = require('../data/factor-enroll-password.json');
+//const data1 = require('../data/factor-enroll-password.json');
 // const data1 = require('../data/factor-enroll-options.json');
 // const data1 = require('../data/factor-verification-email.json');
 // const data1 = require('../data/terminal-email-return.json');
